api/requirements.txt
flask==2.0.1
flask-cors==3.0.10
sentence-transformers==2.2.2
faiss-cpu==1.7.3
numpy==1.22.0
pandas==1.4.2
